Move non-public modules to nglient/_internal
Separate public/private API. Keep modules
containing the piblic classes in the main client
directory and move the rest to _internal.
