+++ tuf/ngclient/_internal/download.py
