+++ tuf/ngclient/_internal/metadata_bundle.py
